Refactor import statements and implement dungeon factory pattern for maze generation
