+++ tests/test_dungeon.py
@@ -5,7 +5,8 @@ import os
 # Add the src directory to the system path
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../src')))
 
-from Dungeon import Dungeon
+from Adventurer import Adventurer
+from dungeon.Dungeon import Dungeon
 
 # test/test_Dungeon.test.py
