+++ tests/test_room.py
@@ -4,7 +4,7 @@ import os
 
 # Add the src directory to the system path
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../src')))
-from Room import Room
+from Dungeon.Room import Room
 from Adventurer import Adventurer
 
 class TestRoom(unittest.TestCase):
